Nexţ changes and small code refactoring
