+++ src/components/atoms/CardImage/index.types.ts
@@ -3,4 +3,5 @@ export interface CardImageProps {
   reversed?: boolean;
   onlyMobile?: boolean;
   onlyDesktop?: boolean;
+  bigger?: boolean;
 }
