+++ src/components/atoms/WhenAndWhere/index.module.scss
@@ -8,4 +8,13 @@
     text-align: center;
     font-weight: 600;
   }
+  a {
+    text-decoration: none;
+    color: var(--nt-orange);
+    display: flex;
+    flex-direction: row;
+    gap: 20px;
+    align-items: center;
+    justify-content: center;
+  }
 }
