+++ src/components/organisms/HeroHome/index.tsx
@@ -7,7 +7,7 @@ import Colli from '../../../assets/colli.svg';
 import Collina from '../../../assets/collina.svg';
 import FastActions from '../../molecules/FastActions';
 import Flower from '../../atoms/Flower';
-import { showPopUp } from '../../../hooks/newsletter';
+import { showPopUp } from '../../../utilities/newsletter';
 import SocialLinks from '../../molecules/SocialLinks';
 
 const Index = () => {
